Rewrote plot gathering from files, refactoring
Rewrote how the plots are obtained from the files. Directory structure
is not important anymore, just scan the file recursively and gather all
objects. If more plots of a certain type are available, use the first
one and print a warning.
Refactored the code, program logic should be better now.